Decompile no4 water effects functions (#2627)
Sorry for the massive PR. My original plan was to do this on PSP and PSX
side and have splits and data imported. Main reason I'm putting this up
now is I realised some of what I was working on was starting to overlap
with @cram0's research here and I want to get out of their way as they
seem more on top of what the functions are doing and documenting them
better 😄

As I went along I kept expecting a PSP file split that never came. It
looks like most if not all of the functions in first_c_file are
consistent with what's on the PSP side, so this may just be one big file
(or researching the functions may give some more clear context wrt
splits).

All of these functions should be matching on PSP side also so should be
pretty trivial to codeshare later on.

+++ include/game.h
@@ -1599,7 +1599,7 @@ typedef struct {
     /* 8003C850 */ RelicOrb* relicDefs;
     /* 8003C854 */ void (*InitStatsAndGear)(bool debugMode);
     /* 8003C858 */ s32 (*PlaySfxVolPan)(s32 sfxId, s32 sfxVol, s32 sfxPan);
-    /* 8003C85C */ s32 (*SetVolumeCommand22_23)(s32 vol, s16 distance);
+    /* 8003C85C */ s32 (*SetVolumeCommand22_23)(s32 vol, s32 distance);
     /* 8003C860 */ void (*func_800F53A4)(void);
     /* 8003C864 */ u32 (*CheckEquipmentItemCount)(u32 itemId, u32 equipType);
     /* 8003C868 */ void (*GetPlayerSensor)(Collider* col);
@@ -1680,7 +1680,7 @@ extern void (*g_api_func_800FE044)(s32, s32);
 extern void (*g_api_AddToInventory)(u32 id, EquipKind kind);
 extern RelicOrb* g_api_relicDefs;
 extern s32 (*g_api_PlaySfxVolPan)(s32 sfxId, s32 sfxVol, s32 sfxPan);
-extern s32 (*g_api_SetVolumeCommand22_23)(s32 vol, s16 distance);
+extern s32 (*g_api_SetVolumeCommand22_23)(s32 vol, s32 distance);
 extern void (*g_api_func_800F53A4)(void);
 extern u32 (*g_api_CheckEquipmentItemCount)(u32 itemId, u32 equipType);
 extern void (*g_api_GetPlayerSensor)(Collider* col);
